% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_potential.R
\name{get_statistical_potentials}
\alias{get_statistical_potentials}
\title{Compute Statistical Potentials for Interface Contacts}
\usage{
get_statistical_potentials(df_contacts, parts = "all", id_col = "pdb_id")
}
\arguments{
\item{df_contacts}{A data.table or data.frame containing the contact information with at least
three columns: `resid_ab`, `resid_ag`, and `id_col` (default: 'pdb_id'). This data frame
should already be filtered based on distance threshold (e.g., distance < 8.5).}

\item{parts}{A character vector specifying the parts of the protein for which to calculate the potentials.
Default is 'all', but can also include 'h3', 'l3', or other cdr's loops.}

\item{id_col}{The name of the column to use as the identifier (e.g., 'pdb_id'). Default is 'pdb_id'.}
}
\value{
A data.table containing the summed statistical potentials for each identifier,
  with columns for symmetric (sp_sym_<part>) and asymmetric (sp_asym_<part>) potentials for each part.
  The dataframe is grouped by the specified identifier (e.g., 'pdb_id').
}
\description{
Computes the sum of statistical potentials (both symmetric and asymmetric)
for antibody-antigen interface contacts, based on input data frames.
}
