% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_poses.R
\name{preprocess_structure}
\alias{preprocess_structure}
\title{Preprocess biological structure (pdb)}
\usage{
preprocess_structure(df_structure, id_col)
}
\arguments{
\item{df_structure}{data.frame or data.table with columns:
id_col, elety, resid, resno, insert, chain, x, y, z}

\item{id_col}{character string specifying the column used as frame ID (default "pdb_id")}
}
\value{
data.table merged with force field parameters, ready for energy calculations
}
\description{
This function standardizes atom names, identifies N- and C-terminal atoms,
and merges the pose data with force field parameters.
}
